fix(iter-3): move resolve_sidecar_pair to core, extract anchor types, consolidate save routing
- Move resolve_sidecar_pair to core::paths routing through
  resolve_sidecar_for_file for TOCTOU canonicalization (security fix)
- Add save_sidecar_routed to core::sidecar consolidating
  ensure_parent + save_at pattern (was duplicated 3)
- Extract NewCommentAnchor/TaggedNewAnchor to anchor_input.rs submodule
- commands/comments/mod.rs: 462351 lines (budget: 400)
