+++ src-tauri/src/commands/comments/mod.rs
@@ -1,22 +1,25 @@
 //! Comment thread mutation commands (sidecar reads, writes, anchor hashing).
 //!
-//! Split into 4 submodules for the 400-LOC budget (architecture rule 23):
+//! Split into 5 submodules for the 400-LOC budget (architecture rule 23):
 //! - `mod.rs` — workspace guard + CRUD entry points
+//! - `anchor_input.rs` — `NewCommentAnchor` / `TaggedNewAnchor` wire types
 //! - `badges.rs` — `get_file_badges`
 //! - `get.rs` — `get_file_comments` (typed-anchor dispatch + matching)
 //! - `update.rs` — `update_comment` + `CommentPatch`
 
 use crate::core::mrsf_version::MRSF_VERSION_DEFAULT;
-use crate::core::types::{Anchor, CommentAnchor, MrsfSidecar, WordRangePayload};
-use serde::Deserialize;
+use crate::core::types::{Anchor, MrsfSidecar};
 use std::path::Path;
 use tauri::{AppHandle, Emitter, Runtime, State};
 
 use crate::watcher::{SidecarConfigState, WatcherState};
 
+pub mod anchor_input;
 pub mod badges;
 pub mod get;
 pub mod update;
+
+pub use anchor_input::{NewCommentAnchor, TaggedNewAnchor};
 
 pub use badges::{get_file_badges, get_file_badges_inner, FileBadge};
 pub use get::{get_file_comments, get_file_comments_inner, GetFileCommentsResult};
@@ -77,36 +80,16 @@ impl<R: Runtime> CommentsEmitter for AppHandle<R> {
     }
 }
 
-/// Compute the YAML and JSON sidecar paths for a source file, consulting
-/// the workspace's `.mrsf.yaml` config if present. When `sidecar_root` is
-/// configured, paths are redirected under `workspace_root/sidecar_root/`.
-/// Falls back to co-located paths when no config is active.
-///
-/// Returns `(yaml_path, json_path, Option<workspace_root>)`. The third
-/// element is `Some` when a redirected config is active — callers use it
-/// to call [`crate::core::paths::ensure_sidecar_parent`] before writing.
+/// Thin wrapper around [`crate::core::paths::resolve_sidecar_pair`] that
+/// extracts the workspace config from managed state.
 pub(crate) fn resolve_sidecar_pair(
     file_path: &str,
     config_state: &SidecarConfigState,
 ) -> (String, String, Option<std::path::PathBuf>) {
     let file = Path::new(file_path);
-    if let Some((ws_root, Some(sr))) = config_state.resolve_for_file(file) {
-        if let Ok(relative) = file.strip_prefix(&ws_root) {
-            let sidecar_dir = ws_root.join(&sr);
-            let yaml = sidecar_dir.join(format!("{}.review.yaml", relative.display()));
-            let json = sidecar_dir.join(format!("{}.review.json", relative.display()));
-            return (
-                yaml.to_string_lossy().into_owned(),
-                json.to_string_lossy().into_owned(),
-                Some(ws_root),
-            );
-        }
-    }
-    (
-        format!("{}.review.yaml", file_path),
-        format!("{}.review.json", file_path),
-        None,
-    )
+    let config_data = config_state.resolve_for_file(file);
+    let config = config_data.as_ref().map(|(ws, sr)| (ws.as_path(), sr));
+    crate::core::paths::resolve_sidecar_pair(file, config)
 }
 
 /// Load a sidecar, apply a mutation, save, and emit `comments-changed`.
@@ -121,11 +104,7 @@ fn with_sidecar_mut<E: CommentsEmitter>(
         .map_err(|e| e.to_string())?
         .ok_or("sidecar not found")?;
     mutate(&mut sidecar)?;
-    let save_path = std::path::PathBuf::from(&yaml);
-    if let Some(ref root) = ws_root {
-        crate::core::paths::ensure_sidecar_parent(root, &save_path).map_err(|e| e.to_string())?;
-    }
-    crate::core::sidecar::save_sidecar_at(&save_path, &sidecar.document, &sidecar.comments)
+    crate::core::sidecar::save_sidecar_routed(&yaml, ws_root.as_deref(), &sidecar.document, &sidecar.comments)
         .map_err(|e| e.to_string())?;
     emitter.emit_comments_changed(file_path);
     Ok(())
@@ -156,11 +135,7 @@ pub fn mutate_sidecar_or_create(
             comments: vec![],
         });
     mutate(&mut sidecar)?;
-    let save_path = std::path::PathBuf::from(&yaml);
-    if let Some(ref root) = ws_root {
-        crate::core::paths::ensure_sidecar_parent(root, &save_path).map_err(|e| e.to_string())?;
-    }
-    crate::core::sidecar::save_sidecar_at(&save_path, &sidecar.document, &sidecar.comments)
+    crate::core::sidecar::save_sidecar_routed(&yaml, ws_root.as_deref(), &sidecar.document, &sidecar.comments)
         .map_err(|e| e.to_string())?;
     Ok(())
 }
@@ -182,94 +157,6 @@ fn with_sidecar_or_create<E: CommentsEmitter>(
 // `get_file_comments` lives in [`get`] (split out to keep this file under
 // the architecture rule 23 LOC budget). Re-exported above so the IPC
 // registration in `lib.rs` stays unchanged.
-
-/// Wire-format anchor for `add_comment`. Accepts BOTH the legacy flat
-/// `{ line, ... }` shape used by line-anchored composers and the tagged
-/// `{ kind: "...", ... }` shape introduced for file-level + typed
-/// anchors (Group A/B). Untagged so the JS chokepoint (`addComment` in
-/// `lib/tauri-commands.ts`) does not have to convert.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(untagged)]
-pub enum NewCommentAnchor {
-    Tagged(TaggedNewAnchor),
-    Legacy(CommentAnchor),
-}
-
-/// Tagged variant of [`NewCommentAnchor`]. Mirrors the TS `Anchor` union
-/// in `src/types/comments.ts` — discriminator is `kind`, payload fields
-/// are flattened alongside it (internally tagged).
-#[derive(Debug, Clone, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum TaggedNewAnchor {
-    Line {
-        line: u32,
-        #[serde(default)]
-        end_line: Option<u32>,
-        #[serde(default)]
-        start_column: Option<u32>,
-        #[serde(default)]
-        end_column: Option<u32>,
-        #[serde(default)]
-        selected_text: Option<String>,
-        #[serde(default)]
-        selected_text_hash: Option<String>,
-    },
-    File,
-    WordRange(WordRangePayload),
-}
-
-impl NewCommentAnchor {
-    /// Convert into the canonical in-memory [`Anchor`] enum + a legacy
-    /// flat [`CommentAnchor`] (used by `create_comment` to populate the
-    /// MrsfComment's flat line fields). For non-Line variants, the flat
-    /// fields are left as the default — callers must not rely on them.
-    /// Exposed `pub` for integration tests of `add_comment`'s anchor
-    /// dispatch (the `#[tauri::command]` itself can't be invoked outside
-    /// a Tauri runtime).
-    pub fn into_anchor_pair(self) -> (Anchor, Option<CommentAnchor>) {
-        match self {
-            NewCommentAnchor::Legacy(c) => {
-                let anchor = Anchor::Line {
-                    line: c.line,
-                    end_line: c.end_line,
-                    start_column: c.start_column,
-                    end_column: c.end_column,
-                    selected_text: c.selected_text.clone(),
-                    selected_text_hash: c.selected_text_hash.clone(),
-                };
-                (anchor, Some(c))
-            }
-            NewCommentAnchor::Tagged(TaggedNewAnchor::Line {
-                line,
-                end_line,
-                start_column,
-                end_column,
-                selected_text,
-                selected_text_hash,
-            }) => {
-                let flat = CommentAnchor {
-                    line,
-                    end_line,
-                    start_column,
-                    end_column,
-                    selected_text: selected_text.clone(),
-                    selected_text_hash: selected_text_hash.clone(),
-                };
-                let anchor = Anchor::Line {
-                    line,
-                    end_line,
-                    start_column,
-                    end_column,
-                    selected_text,
-                    selected_text_hash,
-                };
-                (anchor, Some(flat))
-            }
-            NewCommentAnchor::Tagged(TaggedNewAnchor::File) => (Anchor::File, None),
-            NewCommentAnchor::Tagged(TaggedNewAnchor::WordRange(p)) => (Anchor::WordRange(p), None),
-        }
-    }
-}
 
 /// Create a new comment, save to sidecar.
 ///
